增加根据前缀查找Object的接口

+++ common/pkgs/storage/s3/shard_store.go
@@ -104,6 +104,10 @@ func (s *ShardStore) removeUnusedTempFiles() {
 		marker = resp.NextMarker
 	}
 
+	if len(deletes) == 0 {
+		return
+	}
+
 	resp, err := s.cli.DeleteObjects(context.Background(), &s3.DeleteObjectsInput{
 		Bucket: aws.String(s.bucket),
 		Delete: &s3types.Delete{
